requirements.txt
# Core ML/RL
torch>=2.1
stable-baselines3>=2.3
gymnasium>=0.29

# NLP / Embeddings
sentence-transformers>=3.0

# Utilities
numpy<2.0
pyyaml
tensorboard
matplotlib

# Testing
pytest
